+++ common/src/main/java/com/st0x0ef/stellaris/client/overlays/JetSuitOverlay.java
@@ -1,7 +1,6 @@
 package com.st0x0ef.stellaris.client.overlays;
 
 import com.st0x0ef.stellaris.Stellaris;
-import com.st0x0ef.stellaris.common.items.armors.AbstractSpaceArmor;
 import com.st0x0ef.stellaris.common.items.armors.JetSuit;
 import com.st0x0ef.stellaris.common.utils.FuelUtils;
 import com.st0x0ef.stellaris.common.utils.Utils;