+++ changelog.md
@@ -11,6 +11,7 @@
 ## Fix
 - Fix fuel refinery doesn't convert convert neo mb to fabric mb
 - Fix fluid transfer
+- Fix a few structures gen
 
 
 # 1.2.0 final changelog